api_requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
Pillow>=8.3.0
hnswlib>=0.7.0
scikit-learn>=1.0.0
ftfy>=6.0.0
regex>=2021.4.4 